+++ src/components/FileSelect/index.tsx
@@ -86,16 +86,14 @@ class FileSelectInt extends React.Component<FileSelectProps & FormContextProps, 
     }
   }
 
-  componentWillReceiveProps(e: FileSelectProps) {
-    const { controlled, value } = this.state;
-
-    if (controlled && e.value && value !== e.value) {
-      this.setState(() => ({
-        value: e.value || [],
+  componentWillReceiveProps({ value = [], error }: FileSelectProps) {
+    if (this.state.controlled && value && this.state.value !== value) {
+      this.setState({
+        value,
         previews: [],
-        error: e.error,
-      }));
+      });
     }
+    this.setState({ error });
   }
 
   private addFileEntries = (ev: React.ChangeEvent<HTMLInputElement>) => {
@@ -228,24 +226,11 @@ class FileSelectInt extends React.Component<FileSelectProps & FormContextProps, 
   };
 
   render() {
-    const {
-      children,
-      theme,
-      value: _0,
-      defaultValue: _1,
-      disabled,
-      multiple,
-      info,
-      onChange: _2,
-      preview: _3,
-      onOpen: _4,
-      onInput: _5,
-      ...props
-    } = this.props;
+    const { children, disabled, multiple, info } = this.props;
     const { value, error } = this.state;
 
     return (
-      <div {...props}>
+      <div>
         <Button onClick={this.openFilePicker} disabled={disabled} buttonStyle="secondary" type="button">
           {children}
         </Button>