+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fixed: missed selection state of `RadioButtonGroup` in non-controlled mode with form context.
 - Compatibility with IE11
 - `Sidebar` component added
+- Fixed: component wrapped with `withValidation` hoc didn't display an error
 
 ## 0.5.0